Major updates with new features:
1. Simulating data stream using Kafka.

2. Rigorous synchronized mode, working together with METS-R HPC.

3. Rewritten communication package:
	a. QUERY and ANSER message handlers/senders.
	b. CONTROL message handlers.

+++ METS_R/src/mets_r/communication/StepMessageHandler.java
@@ -7,7 +7,7 @@ import mets_r.ContextCreator;
 public class StepMessageHandler extends MessageHandler{
 	@Override
 	public String handleMessage(String msgType, JSONObject jsonMsg) {
-		int tick = (int) jsonMsg.get("TICK");
+		int tick = ((Long) jsonMsg.get("TICK")).intValue();
 		if(tick == ContextCreator.getCurrentTick()) {
 			ContextCreator.receivedNextStepCommand = true;
 			return "OK";
